Fix request timeout u64 error (#646)
* fix: switch request ext_timeout to u64

* fix: truncate budget instead of error

* feat: scale query timeout do u64

+++ src/net/primitives.c
@@ -455,7 +455,7 @@ int8_t _z_send_reply_err(const _z_query_t *query, const _z_session_rc_t *zsrc, c
 /*------------------ Query ------------------*/
 int8_t _z_query(_z_session_t *zn, _z_keyexpr_t keyexpr, const char *parameters, const z_query_target_t target,
                 const z_consolidation_mode_t consolidation, _z_value_t value, _z_reply_handler_t callback,
-                _z_drop_handler_t dropper, void *arg, uint32_t timeout_ms, const _z_bytes_t attachment,
+                _z_drop_handler_t dropper, void *arg, uint64_t timeout_ms, const _z_bytes_t attachment,
                 z_congestion_control_t cong_ctrl, z_priority_t priority, _Bool is_express) {
     int8_t ret = _Z_RES_OK;
